--- jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/deploy/SamDeployView.form
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="software.aws.toolkits.jetbrains.services.lambda.deploy.SamDeployView">
-  <grid id="27dc6" binding="content" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="741" height="398"/>
-    </constraints>
-    <properties>
-      <preferredSize width="750" height="350"/>
-      <requestFocusEnabled value="true"/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <nested-form id="a6573" form-file="software/aws/toolkits/jetbrains/ui/ProgressPanel.form" binding="progressPanel" custom-create="true">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-      </nested-form>
-      <tabbedpane id="24f22" class="com.intellij.ui.components.JBTabbedPane" binding="logTabs" custom-create="true">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false">
-            <preferred-size width="200" height="200"/>
-          </grid>
-        </constraints>
-        <properties/>
-        <border type="none"/>
-        <children/>
-      </tabbedpane>
-    </children>
-  </grid>
-</form>
